requirements.txt
fastapi
uvicorn
requests
google-generativeai
kubernetes
tenacity
